Added greater flexibility to JsonWriter
