+++ json/shared/src/main/scala/scribe/json/Trace.scala
@@ -1,0 +1,9 @@
+package scribe.json
+
+import fabric.rw._
+
+case class Trace(message: String, elements: List[TraceElement], cause: Option[Trace])
+
+object Trace {
+  implicit val rw: ReaderWriter[Trace] = ccRW
+}
