+++ json/shared/src/main/scala/scribe/json/TraceElement.scala
@@ -1,0 +1,9 @@
+package scribe.json
+
+import fabric.rw._
+
+case class TraceElement(`class`: String, method: String, line: Int)
+
+object TraceElement {
+  implicit val rw: ReaderWriter[TraceElement] = ccRW
+}
